modify filter dir & version
