+++ common/constant/version.go
@@ -18,7 +18,7 @@
 package constant
 
 const (
-	Version = "2.6.0"
+	Version = "1.3.0"
 	Name    = "dubbogo"
 	DATE    = "2019/05/06"
 )
