+++ filter/filter_impl/access_log_filter_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package default_filters
+package filter_impl
 
 import (
 	"context"
