+++ filter/filter_impl/graceful_shutdown_filter.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package default_filters
+package filter_impl
 
 import (
 	"sync/atomic"
