+++ filter/filter_impl/hystrix_filter_test.go
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package default_filters
+package filter_impl
 
 import (
 	"regexp"
