+++ filter/filter_impl/token_filter_test.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package default_filters
+package filter_impl
 
 import (
 	"net/url"
