+++ filter/filter_impl/tps_limit_filter.go
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 
-package default_filters
+package filter_impl
 
 import (
 	"github.com/apache/dubbo-go/common/constant"
 	"github.com/apache/dubbo-go/common/extension"
 	"github.com/apache/dubbo-go/common/logger"
 	"github.com/apache/dubbo-go/filter"
-	_ "github.com/apache/dubbo-go/filter/default_filters/tps"
+	_ "github.com/apache/dubbo-go/filter/filter_impl/tps"
 	_ "github.com/apache/dubbo-go/filter/handler"
 	"github.com/apache/dubbo-go/protocol"
 )
